[FLINK-19547] Add the partialRecordLength when creating a BufferConsumer
Partial records happen if a record can not fit into one buffer, then the remaining part of the same record
is put into the next buffer. Hence partial records only exist at the beginning of a buffer.
Partial record clean-up is needed in the mode of approximate local recovery.
If a record is spanning over multiple buffers, and the first (several) buffers have got lost due to the failure
of the receiver task, the remaining data belonging to the same record in transition should be cleaned up.

`partialRecordLength` is the length of bytes to skip in order to start with a complete record,
from position index 0 of the underlying MemorySegment. `partialRecordLength` is used in approximate
local recovery to find the start position of a complete record on a BufferConsumer, so-called
`partial record clean-up`.

This commit includes
1). API change to add `partialRecordLength` when creating a BufferConsumer
2). Add `partialRecordLength` in `BufferWritingResultPartition` when emitRecord and `broadcastRecord`